refactor(orchestrator): extract deterministic probability-tutor brain from the God-file (D-163 Stage 1/M13)
Decompose app/infrastructure/clients/orchestrator_client.py (6,154 lines) by lifting
the ~2,600-line deterministic probability-tutor brain — verbatim, zero behavioral change —
into a standalone app/services/skills/probability_tutor_brain.py (ProbabilityTutorBrain).
OrchestratorClient now inherits it as a mixin, so every cls._x / self._x call-site in
chat_with_agent resolves via the MRO unchanged. The God-file shrinks to 3,832 lines (-38%).

The brain is proven transport-free (all @classmethod/@staticmethod over
(question, history, combo, kc_progress); zero HTTP / instance-state coupling). It is a live
consumer via inheritance — not a ZOMBIE (import + call chain + runtime evidence).

Blast radius handled in lockstep (D-013/D-140 pattern — alignment, never disablement):
- scripts/fitness/check_pedagogical_os.py: source-inspection gates read the brain for moved
  def/body pins; check_split_brain_parity now compares brain <-> orchestrator port directly
  and asserts OrchestratorClient still inherits the brain (no-ZOMBIE guard).
- scripts/fitness/check_skills_doctrine.py: DialogueManager wiring split (call-site in client,
  capability in brain).
- 11 source-inspection test files: CLIENT_SRC now unions client + brain so any pin survives;
  two inter-method slices repointed to the next in-brain method.

Verified: 746 passed / 2 deselected (pre-existing event-loop failures already in CI deselect)
across the full 52-file blast radius; runtime import + inheritance + MRO resolution intact;
check_pedagogical_os + check_skills_doctrine + runtime_truth --check + validate_structure +
ci_guardrails + check_test_hygiene + ruff check/format all green.
