+++ tests/services/test_iss126_step_explanation.py
@@ -61,9 +61,12 @@ def _combo() -> SimpleNamespace:
 
 
 def _client_src() -> str:
-    return (_REPO_ROOT / "app" / "infrastructure" / "clients" / "orchestrator_client.py").read_text(
-        encoding="utf-8"
-    )
+    # D-163: عقل الاحتمالات استُخرج للوحدة المستقلة — نضمّ الاثنين ليصمد أي pin.
+    return (
+        _REPO_ROOT / "app" / "infrastructure" / "clients" / "orchestrator_client.py"
+    ).read_text(encoding="utf-8") + (
+        _REPO_ROOT / "app" / "services" / "skills" / "probability_tutor_brain.py"
+    ).read_text(encoding="utf-8")
 
 
 # ─────────────────────────────────────────────────────────────────────────────
