+++ tests/services/test_iss128_question_speech_act.py
@@ -93,9 +93,12 @@ def _other_combo() -> SimpleNamespace:
 
 
 def _client_src() -> str:
-    return (_REPO_ROOT / "app" / "infrastructure" / "clients" / "orchestrator_client.py").read_text(
-        encoding="utf-8"
-    )
+    # D-163: عقل الاحتمالات استُخرج للوحدة المستقلة — نضمّ الاثنين ليصمد أي pin.
+    return (
+        _REPO_ROOT / "app" / "infrastructure" / "clients" / "orchestrator_client.py"
+    ).read_text(encoding="utf-8") + (
+        _REPO_ROOT / "app" / "services" / "skills" / "probability_tutor_brain.py"
+    ).read_text(encoding="utf-8")
 
 
 def _port_src() -> str:
